refactor (removed interaction function to new commander) & test
